pyproject.toml
[tool.poetry]
name = "library-management"
version = "0.1.0"
description = ""
authors = ["Julia Bengert <yubengert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.7.0"
redis = "^5.2.1"
celery = "^5.5.2"
django-celery-beat = "^2.8.0"
gevent = "^25.4.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
isort = "^6.0.1"
black = "^25.1.0"
coverage = "^7.8.0"


[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
